chromeos-auth-config: drop chronos /etc/shadow check
We've moved towards only using the cros_debug signal for checking
dev mode state.  It's what we've documented, and nowhere else do
we check for "chronos" in the /etc/shadow file.  We don't have too
many users of this script (as many places run crossystem directly),
and a cursory check of those suggests they'd be fine with this.

BUG=None
TEST=CQ passes

Change-Id: I3f7b10f661e1155c296a86138c85fb4fac95f303
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1968269
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>