ec-utils: don't install userspace init scripts
This was used by Link to install some userspace scripts for the
lightbar (not used by Samus). We never used this for anything past
Link, and Link is now AUE, so delete this code which depends on the
BOARD for utils.

BUG=chromium:1035536
TEST=emerge-drallion ec-utils

Change-Id: I0635a9d1d005df7680ef9707406fe1afe3fe0ca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1974631
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>