dev-libs/openssl: Downgrade to 1.0.2
The 1.1.1 upgrade broke wpa_supplicant, so revert to 1.0.2 for now.
Note that this keeps the existing slot structure, but changes the
ebuilds for 1.0.2 to install the package artifacts and 1.1.1 to only
drop the shared objects in the chroot (the inverse of what the uprev
did).

BUG=b:146758562
TEST=Builds

Change-Id: I80387144e5df37fa0c0b57640a3acee86ccb3e1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1980572
Tested-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Commit-Queue: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>