Reduce libassistant.so size
A recent change to the Assistant integration tests increased
the size of the libassistant.so library with 1.3M.
Because this change is only required for testing purposed,
a flag was introduced so we can keep the library size down for
the release build.

This change enables this flag.

Library size after this change:
$ ls -l /tmp/m/./opt/google/chrome/libassistant.so
-rwxr-xr-x 1 root root 11853344 Jan 13 17:45 /tmp/m/./opt/google/chrome/libassistant.so

Depends-On: crrev.com/c/1981131, crrev.com/i/2353499, crrev.com/c/1992346
BUG=b:146653931
TEST=Compiled and deployed
Change-Id: I32d66a34152b6d58df5aee2087135a2feb339f1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1981080
Tested-by: Jeroen Dhollander <jeroendh@google.com>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>