Revert "Adding ability to set USE=cifs to kernel"
This reverts commit 0dc2b97f2c9a1c617744408a6b59569bab4d2773.

Commit 0dc2b97f2c9a ("Adding ability to set USE=cifs to kernel") was added
to support Quanta's automation process. Later it turned out that it
wasn't needed, and the bug was marked WontFix. Nevertheless, the cifs entry
in the eclass file was never removed. Remove it now to clarify that CIFS
is not used in Chrome OS, if for nothing else to avoid having to track
down its non-use whenever security issues are reported against CIFS.

BUG=b:35508715, chromium:1035820
TEST=None

Change-Id: If58900de73bfc7bf43894ae6c54bd4e1bfea62b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1985228
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>