cr50-set-board-id: whitelabel vs mp flag difference is ok
The new whitelabel phase and flags have been added to cr50 and the
cr50-set-board-id scripts. Some boards have been using mp as the phase,
but are going to switch to whitelabel. This means old boards may have mp
flags when the new factory scripts expect whitelabel. It is not a big
deal if old boards have mp flags instead of whitelabel. This change
modifies the error, so factory scripts can ignore the difference in
board id.

This changes cr50_check_board_id_and_flag to fail with ERR_ALREADY_SET
if the only difference between the expected and set board id is 0x4000
in the flags. This bit is the only difference between whitelabel and mp.
Factory scripts consider ERR_ALREADY_SET to be an acceptable error, so
this change still makes cr50-set-board-id fail with an error, but it's
one that the factory will ignore.

BUG=b:133159485
TEST=set flags to 0x3f80. Try to run 'cr50-set-board-id mp'. Make sure
the script fails with exit code 2. Set flags to 0x7f80. Run
'cr50-set-board-id whitelabel'. Make sure it fails with exit code 2.

Change-Id: I6a490b7bf32053934316fe189ea08747bffe30bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1992053
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>