selinux-policy: Switch to permissive on 5.4 kernel
SELinux in the 5.4 kernel is not compatible with the current Android
sepolicy, which prevents ARC++ from booting.  The long-term solution
will be to update the policy, but disable selinux until this is done to
unblock testing on moblab, etc.

There should be no danger of accidentally shipping a device configured
this way, because we can't pass CTS until selinux is turned back on.

BUG=b:144691707
TEST=emerge-grunt-kernelnext selinux-policy shows permissive
     emerge-octopus selinux-policy still shows enforcing

Change-Id: I224251d65cc755ca70fa04e56f8fde93a4bc5bd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1990066
Reviewed-by: Qijiang Fan <fqj@google.com>
Reviewed-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>