lld: Turning on --no-rosegment by default.
--no-rosegment flag is a linker flag to keep read-only data in the
same segment of executables, and the default behavior of LLD is to
enable rosegment. This patch tries to keep no-rosegment as the
default behavior of LLD, to match the gold linker, as well as
keeping symbolization downstream tools easy to maintain. The
experiments show no performance difference with and without
--no-rosegment, and we see a reduction of ~50kB size with this
patch on ~5 boards we tested.

BUG=b:140622025
TEST=Nightly tests and rotating builders by toolchain team.

Change-Id: Ie3b743576fb936a09f4f411f1085e1168ecb69c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1961032
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>