cros-workon: fix a bug in checking if CROS_WORKON_PROJECT is empty.
After running array_vars_autocomplete, even if the input
CROS_WORKON_PROJECT is empty, it may become a list of empty strings.
Then it is not right to check its emptiness by

[[ -n "${CROS_WORKON_PROJECT[*]}" ]]

because ${CROS_WORKON_PROJECT[*]} will be a string of blanks and
non-empty.

BUG=none
TEST=on workstation, "emerge-eve ml" works.

Change-Id: I0535ed0b4c63c34eaf692adc958605ed1ab2e5d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2003988
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>