toolchain-packages: Remove dev-lang/rust from special CQ coverage
dev-lang/rust requires every cross-compiler to be built in order to be
compiled from source. The CQ builders only build the toolchains they
need for the particular build target they run, so dev-lang/rust is
currently not able to be tested in the same manner as the rest of the
toolchain packages.

BUG=None
TEST=None

Change-Id: I8b2b77b00f90472d3978b36b8cd9528bb09d44b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2005066
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>