chromeos-base/emright_fw_updater: fix license
The license in this ebuild was incorrectly listed as "GPL-2.0", which
caused builders to fail once the package was built:
  https://logs.chromium.org/logs/chromeos/buildbucket/cr-buildbucket.appspot.com/8890646680779685056/+/steps/build_images/0/steps/call_chromite.api.ImageService_Create/0/steps/call_build_API_script/0/stdout

This commit changes the license string to GPL-2.

This commit does not update the revision number, as it fixes a
build-time issues with licenses.
https://devmanual.gentoo.org/general-concepts/ebuild-revisions/index.html

BUG=b:141912132
TEST=None

Change-Id: I1a4c60ed5b880e35af379ebdd1f329f2def93d98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2012590
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Justin TerAvest <teravest@chromium.org>
Tested-by: Justin TerAvest <teravest@chromium.org>