dev-libs/libgpiod: Mask C++ library bindings
The 1.4.1 libgpiod ebuild enables cxx library bindings via the 'cxx' use
flag. This fails to compile in two ways. First it uses exceptions, so
we enable exceptions explicitly here. Second it fails to link
libgpiodcxx so we just mask away the library because we're not using it.

Given that we disable C++ bindings entirely it is not too useful to
enable exceptions for them in the chromeos profile, but it's done here
so that future developers can enable C++ bindings in the future if
they're interested in figuring out the link problems.

BUG=None
TEST=emerge-trogdor -j libgpiod

Change-Id: I7e295284e1c22fea6b309947ae516c946224b1d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2012593
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>