chromeos-base/arc-keymaster: Add dependencies and update context patch
Keymaster now encrypts the keys of Android applications. For that we
depend on chaps, session manager, cryptohome, and protobuf.

The context is now implemented in a subclass in platform2, so the
context patch no longer needs to be applied. We still do a dry run to
verify overridden functions haven't changed. This guard is in place
because we may need to mirror some of the upstream changes back into
platform2.

BUG=b:139142065, chromium:1040635
TEST=emerge-$BOARD chromeos-base/arc-keymaster

Cq-Depend: chromium:2048063
Change-Id: I60c750c6a662360fa82f19a5130d9b74b5e08edc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2050600
Tested-by: Edman Anjos <edman@chromium.org>
Commit-Queue: Edman Anjos <edman@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>