arc: add dependency between arc-sdcard and arc-setup
arc-sdcard.conf is sourced from multiple paths: either from arc-sdcard
(for the sdcard daemon) or arc-setup (esdfs). On enabling/disabling the
USE flag the first time, this may lead to file collisions if the
packages are emerged in a certain order. Add a dependency to emerge
the package that doesn't provide the file first in both cases.

BUG=chromium:1045176
TEST=emerge with and without USE flag; no conflicts.

Change-Id: I58b1652911fc43d7aff34d40b1c25b69996fbdcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2018326
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>