net-print/ippusbxd: Adding keep-alive patch
This change updates ippusbxd to support the use of 'keep-alive' messages
on a dedicated socket. This message will help ippusb_manager notify
ippusbxd that it wants to begin another CUPS connection. If a
'keep-alive' message is successfully received from ippusb_manager then
the 10-second idle countdown will be reset in order to accommodate the
new request, and it will send ippusb_manager and acknowledgement that is
has done so.

This change also makes a handful of minor changes in order eliminate the
possibility of race conditions or timeouts when using this protocol. The
UDS socket which is used to communicate with CUPS is now the final thing
that is closed before shutting down. This is done so that ippusb_manager
will know when it is safe to initiate a new ippusbxd process.

There was also an unnecessary call to libusb_reset_device() in the USB
shutdown code which was causing the actual shutdown process of ippusbxd
to take a painfully long time. Since the USB interfaces are being
released anyways, there's no point in attempting to perform this lengthy
reset. As a benefit, the removal of this reset has improved the
robustness of the Virtual USB Printer which currently does not support
handling these requests.

BUG=chromium:1047022
TEST=Tested manually using virtual-usb-printer

Cq-Depend: chromium:2032705
Change-Id: Iceac42769265c79640fa1e7d36f1e87ce6c07e7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2032665
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: David Valleau <valleau@chromium.org>
Auto-Submit: David Valleau <valleau@chromium.org>