volteer: Add ripto coreboot configuration
Create a new coreboot configuration for the ripto variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.0.0).

BUG=b:147856356,b:150747922
TEST=FW_NAME=ripto emerge-volteer coreboot chromeos-bootimage
Ensure that image-ripto.*.bin are created

Reviewed-on: https://chrome-internal-review.googlesource.com/c/chromeos/overlays/chipset-tgl-private/+/2469850
Commit-Queue: Paul Fagerburg <pfagerburg@google.com>
Tested-by: Paul Fagerburg <pfagerburg@google.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Change-Id: Ic19a49257083766566cc2768a2b27fcade2559de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2087156
Reviewed-by: caveh jalali <caveh@chromium.org>
Commit-Queue: caveh jalali <caveh@chromium.org>
Tested-by: caveh jalali <caveh@chromium.org>