dev-lang/ghc: Set compiler path when building
The src_configure stage expects to be able to run ghc to determine its
supported build flags.  If ghc isn't already installed, we need to
explicitly set HC to the path of the bootstrap ghc so that this won't
attempt to run "--info" as a command.

BUG=chromium:1028204
TEST=sudo emerge shellcheck; chromiumos-sdk tryjob

Change-Id: I74e43c639d6e8df01caf8d1d4a9b3053c18f3a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2027946
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>