chromeos-chrome: Build Dawn test binaries again
Once upon a time, we were building the Dawn unit and end-to-end test
binaries for use in chrome-binary-tests [1]. These were later
removed as a dependency change caused compilation issues [2].
Recently, a new build_dawn_tests flag was added [3] so we can build
the test binaries for Chrome OS again independent of including Dawn
in the Chrome compilation.

This CL re-enables the compilation of dawn_unittests and
dawn_end2end_tests as part of chromeos-chrome now that we have
build_dawn_tests. Note that we will still need a follow-up CL in
chrome-binary-tests to actually copy the binaries to the DUT.

[1] https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1757308
[2] https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1919689
[3] https://chromium-review.googlesource.com/c/chromium/src/+/2028649

BUG=chromium:993457,chromium:1042442

Change-Id: I1ba9b56ce0b8ade89b4c3b16dca61c238041b681
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2036672
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>