chrome-binary-tests: Put dawn tests on test image again
Previously, we built the Dawn unit/end-to-end test binaries and
installed them to the to device image in chrome-binary-tests [1].
Unfortunately, we had to revert due to a compilation issue [2], but
because of a recent change to the way Dawn is built [3], we can once
again run these tests.

Note: This cannot be submitted until the corresponding CL in
chromeos-chrome [4] has landed and is included a Chrome uprev.

[1] https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1787882
[2] https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1919689
[3] https://chromium-review.googlesource.com/c/chromium/src/+/2028649
[4] https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2036672

BUG=chromium:993457,chromium:1042442

Change-Id: Ic55fc9b0761e05d09b17770c2b5c35238c5f48ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2036673
Tested-by: Brian Ho <hob@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>