media-libs/minigbm: disable incremental builds
When doing incremental builds of minigbm through emerge, changes to
header files do not rebuild .o files. As running make manually in
the minigbm directory does properly cause header changes to trigger
rebuilds, this seems to be an issue with the portage integration.

Rather than try to decipher the build system, just disable incremental
builds for minigbm. As minigbm is a relatively small library, the cost
of a full rebuild is dwarfed by the overhead of running emerge.

TEST=check that .o files are rebuilt after touching a header
BUG=chromium:1048975

Change-Id: I9d2d4194613426f84fda3b18fdc62ee98fa4b7eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2038414
Tested-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>