Retrieve a specific version of cloud_policy.proto in ChromeOS
Previously package chromeos-chrome was always being updated to the
latest version of Chrome, and always installing the newest version
of cloud_policy.proto. That had several problems recently, and in
order to make it consistent with all of the other protofiles; it
was required to retrieve a specific version of cloud_policy.proto.
That version should be consistent with the other protobuf files
version, which are being retrieved in ChromeOS.

This CL is doing the following:
- Retrieving a specific version of cloud_policy.proto, which is the
same version as other protobuf files, which exist in Chromium side
- Making all protobuf files being retrieved from one package only
which is protofiles
- Making a workaround regarding changing the package which installs
the cloud_policy protobuf file, in order to avoid errors from portage
regarding the same file is being installed by both packages. This
renaming workaround is being used until the PUpr uprev the new
version of chromeos-chrome package.

BUG=chromium:1025188, chromium:1009436, chromium:1025717
TEST=emerge --board=${BOARD} chromeos-base/chromeos-chrome protofiles

Cq-Depend: chromium:1929052
Change-Id: I398215d1db436fa8eeb4e7d74078e0234eedc126
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2043844
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Amr Aboelkher <amraboelkher@google.com>
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>