vboot: install all objects even when USE=fuzzer
My previous assumption about USE=fuzzer was that it is a special
mode in which the usual objects compiled and installed by the
ebuild are not needed.

It turns out that some packages depend on having libvboot_host.a
available, even when USE=fuzzer is set.  Specifically,
chromeos-base/metrics is now failing to build when USE=fuzzer is
set for both of these packages.

Since we don't know what is needed and what is not by other
packages when USE=fuzzer is set, just unconditionally compile and
install all vboot_reference objects.

BUG=b:124141368, chromium:605348, chromium:1039772
TEST=USE=fuzzer emerge-eve vboot-reference chromeos-base/metrics
TEST=make clean && make runtests
BRANCH=none

Change-Id: I723955e438d94abab7b9e822df0cbc8c773f5cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2046492
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Commit-Queue: Joel Kitching <kitching@chromium.org>