update-utils: Remove stateful_update
We don't need stateful_update on the device anymore. It has been
chromitized.

Also we never needed quick-provision on the DUT. We always get the ToT
one for provisioning purposes. So we don't need to put it on the DUT.

In addition, we never get to use nebraska.py as a library on DUT. So
remove that part too.

BUG=chromium:1026933
TEST=CQ passes

Cq-Depend: chromium:1963124
Change-Id: I8d90ca783281beedf69ccb9f368d130430c69d3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2054077
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>