virtual/target-chromium-os-sdk: Remove matplotlib
CL:1151530 added matplotlib to the chroot so that certain autotests
could generate graphs. This was never merged into the tree; even if
this functionality were to be added later, it would likely be an
addition to tast rather than autotest. Might as well remove this since
it's not being used.

BUG=chromium:867070
TEST=CQ

Change-Id: Ic9ce381a806fb4b7fdc17b39d33852cca8bb864c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2053587
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>