qualcomm_defconfig: Make CONFIG_USB=y
This config is getting set to be =m now, and that is causing dwc3 to
force gadget mode on all devices on kernels newer than 5.4. Let's make
this explicitly be builtin so that USB drivers like DWC3 that are also
builtin can be configured for dual role support instead of host or
gadget.

BUG=None
TEST=Boot up v5.6-rc1 kernel on cheza. See lsusb output something

Change-Id: Ic3ce4dd9e5fb40446804fe3555bfb9fad752cbfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2056067
Tested-by: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>