platform.eclass: Add use flag to generate compile commands
Add a use flag to generate the compile commands for a platform package.
This creates both _chroot and _no_chroot variants of the file, so that
users can choose the appropriate one based on their editor.

The function itself is a slightly modified version of the one in the
cros-kernel2 eclass.

This also adds a `platform_install_compilation_database` function that
will install the compile_commands files into
/build/compilation_database/${CATEGORY}/${PN}. However ebuilds will need
to manually opt in to this behavior by calling the function.

BUG=none
TEST=`USE=compilation_database emerge-$BOARD vm_host_tools`

Change-Id: I0a4b17cccfaf60d96d2d4093c77c19c161ad220f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2056504
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>