Crostini Tast: Add use flags for revised stable dep.
The crostini_stable dep is now defined in terms of a blacklist of boards
that are not stable, unfortunately there are still quite a few of them.

Keeping the list of flags separate to make it easier to clean them up as
we fix the problems preventing the boards from being stable.

Also removed the atlas flag that was introduced for the crostini_stable
dependency, but is no longer needed.  eve was introduced at the same
time, but is since used by other dependencies as well.

BUG=None
TEST=Verified Eve board runs stable test and skips unstable test.

Change-Id: I7a35d2462f97688bd5a33210235ccbdeb8bc7870
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2064090
Tested-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>
Auto-Submit: Matthew Warton <mwarton@google.com>