Retrieve all protobuf files from protofiles package
There was a workaround that has been done, in order to be able to
install cloud_policy.proto from protofiles package, while its being
installed from chromeos-chrome, just to avoid portage error that
this file is being installed by two packages.
Currently the chromeos-chrome package has picked up the changes in
chromeos-chrome-9999.ebuild to the release version, which enables
us to remove the workaround and retrieve the cloud_policy as well
from protofiles package. In that case all protobuf files are being
installed to a specific version in protofiles package.
This CL will also make a syslink to the *r1.ebuild version of the
protofiles.

BUG=chromium:1054477, chromium:1025188, chromium:1009436, chromium:1025717
TEST=emerge --board=${BOARD} chromeos-base/chromeos-chrome protofiles

Change-Id: I77013e09e0cb5c197452f17281aff11252f714bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2066699
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Amr Aboelkher <amraboelkher@google.com>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>