move Cr50 related stuff from ec-utils to chromeos-cr50-dev
Cr50 tools should stay in sync with the cr50_stab ec tree branch, not
with master.

Let's move building and installing these tools from ec-utils into
chromeos-cr50-dev. Since now this ebuild is going to be both
installing the Cr50 artifacts and Cr50 tools, it needs to be aware if
it is being used for reef, which is the only board which requires the
artifacts.

Making chromeos-cr50-dev a dependency of chromeos-cr50 makes sure that
the Cr50 tools are installed everywhere where Cr50 image is installed,
no need in checking the use flags.

Special care is taken to resolve file collisions.

BUG=b:149350081

TEST=tried running various combinations of emerge, will run tryjob
     builds for reef and octopus (to verify inclusion of artifacts
     only when necessary)

Cq-Depend: chrome-internal:2598885, chromium:2067294, chromium:2067163
Change-Id: I4e4161f802d2690b0a7ea09d5a2b75c2edbf3bc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2067967
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>