Update Googletts to 13.1
This update includes a single change to remove
incognito: "split"
from
manifest_guest.json.

This addition in v13 caused the vent page to be evicted prematurely and for two background pages to be loaded.

Note that this is effectively a revert of v13 back to v12 for the manifest_guest.json file.

BUG=chromium:1044209
TEST=manual; on-device; in OOBE.

Change-Id: Ic781424a9e5ffe3b44ef95099b574b9c8d09feb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2073480
Tested-by: David Tseng <dtseng@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>