icu: change to install static library and necessary headers.
Due to security concerns, we can not let chrome use libicu.so. So we
do not need to compile libicu.so into shared libraries for the
moment. This is not only simpler but also makes the library
smaller (because libtextclassifier is the only user). If more programs
need libicu in the future, we can make it into shared libraries.

This commit also changes to install the necessary header files, rather
than all the header files in chrome/icu.

BUG=chromium:1020393
TEST=works on workstation (emerge-eve icu && emerge-eve ml)

Change-Id: I858bb2f0f32c5287c65af9ee913b21627af078cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2073737
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>