termina_container_tools: add libnss_dns to deps
This library was dlopened and is not available in stretch, so we need to
add the dep manually.

Without it garcon filas to runtime link when it is enabling arc
sideloading and crashes.

BUG=chromium:1057496
TEST=emerge-tatl termina_container_tools

Change-Id: I8540650c48fdfaf88eb96821098037d99d491949
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2081643
Commit-Queue: Nic Hollingum <hollingum@google.com>
Commit-Queue: Stephen Barber <smbarber@chromium.org>
Tested-by: Nic Hollingum <hollingum@google.com>
Auto-Submit: Nic Hollingum <hollingum@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>