perf: Add v5.3.7 base ebuild
This perf version provides better support for symbolication than the
current version.

The ebuild file is a copy from Gentoo and does not work properly as is.
To make it easier to track necessary changes, we modify the ebuild file
in a separate CL.

BUG=b:150829898
TEST=none

Change-Id: I59f7a73d558ee7a6c141a202848fc9c18eb6d518
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2087597
Tested-by: Matthias Springer <springerm@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Matthias Springer <springerm@chromium.org>