perf: Adapt v5.3.7 to work with CrOS
The current v5.3.7 ebuild is copied from Gentoo. This CL modifies the
ebuild file so that it works with CrOS and re-applies patches that were
developed for older perf versions.

As of now, perf v5.3.7 is always built with v4.4 kernel headers. If
features are needed that are only available in newer kernel versions,
the dependency can be adapted after creating the corresponding
linux-headers ebuild.

BUG=b:150829898
TEST=emerge-atlas perf
  verify that v5.3.7 is built
  cros deploy $DUT perf
  run 'perf test'
  => no regressions wrt v4.14

Change-Id: Ie685922b8bceba936751cf25333983e8db4ac026
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2087599
Tested-by: Matthias Springer <springerm@chromium.org>
Commit-Queue: Matthias Springer <springerm@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>