chromeos-chrome: ignore use goma_thinlto if !goma || !thinlto
The goma_thinlto USE flag was introduced so that we can roll out
distributed ThinLTO incrementally and avoid breaking things during
the rollout. In the same spirit of avoiding breakage, this changes
goma_thinlto to only take effect in builds that use Goma and ThinLTO
and making it a no-op otherwise.

R=gbiv,manojgupta
BUG=chromium:877722
TEST=Emerged with a variety of goma, -goma, goma_thinlto, etc.

Change-Id: Ieedd8dd03cfeda09b9e5e612fbbbf19dc7576eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2090577
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>