cros-kernel2: Compile 32-bit vdso on arm64 boards
We must have CROSS_COMPILE_COMPAT and CC_COMPAT in the environment to
compile the 32-bit compatible vdso on arm64 kernels. If the ARCH is
"arm", meaning 32-bit userspace, but the kernel is "arm64" then we
should specify CROSS_COMPILE_COMPAT to be the 32-bit armv7 compiler
prefix and CC_COMPAT to be the path to the 32-bit armv7 compiler (which
is always clang). Check for this condition and pass this on the
commandline so that we can build the 32-bit vdso.

Kernels older than v5.3 won't use these variables so this is safe to
pass in the environment when building older kernel versions.

Primes updated with:

sed -i '/coolest prime/s:131:137:' sys-kernel/chromeos*/*9999.ebuild

BUG=b:151181951, chromium:1060855
TEST=Compile and boot Trogdor, grep vdso /proc/1/maps

Cq-Depend: chromium:2099715
Change-Id: I05be80c294751870ab812cb6ba82595e252956f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2099627
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>