bluez: Support bluez-next USE flag
To support choosing between current and next versions, two cros-workon
projects are declared. During emerge, both project sources are copied to
their respective destination directories, and one is chosen as the
"working directory" in src_prepare() below based on bluez-next USE flag.

BUG=b:150851302
TEST=Test both:
     $ emerge-$BOARD bluez
     $ USE=bluez-next emerge-$BOARD bluez
It's expected that with USE=bluez-next the build still fails because it
needs to be adjusted but better in a separate patch.

Change-Id: If83c13e2a25c5fe026f49a208a9cd0eb799f6ddd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2103480
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>