media-libs/{arc-,}mesa-amd: Work around memory leak.
Fixing it fairly invasive for a cherry-pick to R82, so hence a
workaround.

In the new minigbm code we end up querying the modifier, which triggers
KMS handle management code in radeonsi that leaks the handles. So fully
avoid requesting any KMS handles. (The handles we request in minigbm are
file descriptors, not KMS handles).

BUG=b:151392809
TEST=Run CtsMediaTestCases on Grunt without memory issues.

Change-Id: I5665511678a5d47a6d6f83ff175ef97f9cdea9ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2106835
Tested-by: Bas Nieuwenhuizen <basni@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Commit-Queue: Bas Nieuwenhuizen <basni@chromium.org>