cros-kernel2: restrict binchecks
Adds "binchecks" to RESTRICT for cros-kernel2 so that the build is not
subject to checking for .GCC.command.line, -Wl,-z,relro and -Wl,-z,now.
We don't plan to configure the kernel like that, so those messages just
cause log spam. This change stops the messages from being generated.

BUG=chromium:1061666
TEST=emerge-nocturne --nodeps -e chromeos-kernel-4_4;
  check that GCC.command.line complaints are gone.

Change-Id: I998a3df374ae828b9e8eb076aa2b21bed75ac7fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2106422
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>