chromeos-config-tools: Re-enable converter test
The converter depends on the config repo for proto (API) bindings.

Previously, this required cros-config-api (an sdk dep) in order to use
the bindings, which is unnecessary and makes API changes and local
development more difficult.

This leverages the cros_workon architecture and portage to setup the
local src tree correctly, so the config repo and bindings are
consistently available in the local dev, chroot dev, and portage build
environments.

BUG=None
TEST=via portage emerge, run_tests in the chroot, and outside the chroot

Cq-Depend: chromium:2109501
Change-Id: I4d9862007fd9d04f39ad1385e3f77a0ad3faedf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2109993
Tested-by: C Shapiro <shapiroc@chromium.org>
Auto-Submit: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>