Move the check for BoardID being set after checking for SN Bits
This gives better messages overall to the user. As it is, one BoardID
has been set the user will only see a message about that. Instead we
do still want to tell them about SN Bits having been set to the proper
or a different value.

We only check for BoardID's value if we get an update error from
gcstool. We still show the gsctool error and we check BoardID and
improve the message as a reminder that the failure may be due to
BoardID being set. This allows us to help the user without restricting
their ability to use custom firmwares with different checks.

An example of such an error message (with SN Bits truncated for
readability) is:

> /usr/share/cros/cr50-set-sn-bits.sh
Error 7 while sending vendor command
ERROR: Failed to set SN Bits to 26...b254 (BoardID is set).
>

BUG=chromium:1058432
TEST=run the script

Change-Id: Ia122bba891d4185a076b0b9f81856cb9b1635b9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2110546
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Auto-Submit: Yves Arrouye <drcrash@chromium.org>