chromeos-chrome: explicitly specify use_thin_lto and is_cfi
Recent changes in the chromium build caused it to assert when
use_goma_thin_lto is true, but use_thin_lto isn't. This assert was
being triggered in builds with no explicitly specified value for
use_thin_lto, which we expected to default to true. To make sure
it is in fact true, set it explicitly. Also explicitly specify
is_cfi to prevent it from assuming unexpected values.

BUG=chromium:1062713
TEST=assert no longer happens, use_thin_lto and is_cfi have expected values

Change-Id: Ifc0b758131a8fc69b7f96b6d03d13eb3e43a7f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2111351
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>